feat: refactor nft sale widget component
